Avoid buffer overflows by using long for result

+++ client/shared/src/main/java/it/mulders/modernrpc/client/shared/FibonacciClientOutput.java
@@ -1,4 +1,4 @@
 package it.mulders.modernrpc.client.shared;
 
-public record FibonacciClientOutput(int result, long serverDuration, long clientDuration) {
+public record FibonacciClientOutput(long result, long serverDuration, long clientDuration) {
 }
