+++ shared/core/src/main/java/it/mulders/modernrpc/shared/core/FibonacciOutput.java
@@ -1,4 +1,4 @@
 package it.mulders.modernrpc.shared.core;
 
-public record FibonacciOutput(int result, long duration) {
+public record FibonacciOutput(long result, long duration) {
 }
